==> notebooks/1.IRA_data_preprocessing.ipynb <==
import pandas as pd
import os

folder_path = "../data/raw/2024"

dataframes = []

# Iterate over each file in the folder
for file in os.listdir(folder_path):
    if file.endswith(".csv"):  # Ensure only CSV files are processed
        file_path = os.path.join(folder_path, file)
        try:
            # Try reading with default utf-8 encoding
            df = pd.read_csv(file_path, encoding='utf-8')
        except UnicodeDecodeError:
            # Fallback to latin1 encoding if utf-8 fails
            df = pd.read_csv(file_path, encoding='latin1')
        dataframes.append(df)  # Add DataFrame to the list

# Concatenate all DataFrames into one
combined_df = pd.concat(dataframes, ignore_index=True)

output_file = "../data/processed/2024/combined_2024.csv"
combined_df.to_csv(output_file, index=False, encoding='utf-8')

print(f"Combined CSV saved to {output_file}")


combined_df

combined_df.info()

combined_df.describe()

combined_df.columns

combined_df.drop(["A}äe_nacimiento"], axis = 1, inplace=True)
combined_df.head()

# Rename the columns to standar English
combined_df.rename(columns={
    'Viaje_Id': 'Trip_Id',
    'Usuario_Id': 'User_Id',
    'Genero': 'Gender',
    'Año_de_nacimiento': 'Year_of_Birth',
    'Inicio_del_viaje': 'Trip_Start',
    'Fin_del_viaje': 'Trip_End',
    'Origen_Id': 'Origin_Id',
    'Destino_Id': 'Destination_Id',
}, inplace=True)

combined_df.head()

combined_df.info()

combined_df['Trip_Start'] = pd.to_datetime(combined_df['Trip_Start'], errors='coerce')
combined_df['Trip_End'] = pd.to_datetime(combined_df['Trip_End'], errors='coerce')

combined_df.info()

nat_count = combined_df['Trip_Start'].isna().sum()
print(f"The number of NaT values in the 'Trip_Start' column is: {nat_count}")

nat_count = combined_df.isna().sum()
print(f"The number of NaN values is: {nat_count}")

combined_df['Trip_Start'] = combined_df['Trip_Start'].fillna(method='ffill')
combined_df['Trip_End'] = combined_df['Trip_End'].fillna(method='ffill')

import numpy as np


median_year = combined_df['Year_of_Birth'].median()
combined_df['Year_of_Birth'] = combined_df['Year_of_Birth'].fillna(median_year)

# Get the proportion of M and F
gender_distribution = combined_df['Gender'].value_counts(normalize=True)

# Randomly assign M or F based on the proportion
combined_df['Gender'] = combined_df['Gender'].apply(
    lambda x: np.random.choice(['M', 'F'], p=[gender_distribution['M'], gender_distribution['F']]) if pd.isna(x) else x
)

combined_df["Gender"] = combined_df["Gender"].map({'M': 1, 'F': 0})

nat_count = combined_df.isna().sum()
print(f"The number of NaN values is: {nat_count}")

combined_df['Trip_End'] = pd.to_datetime(combined_df['Trip_End'], errors='coerce')  

# Check for rows with invalid dates (optional)
invalid_dates = combined_df[combined_df['Trip_End'].isna()]
if not invalid_dates.empty:
    print(f"Found {len(invalid_dates)} rows with invalid dates")

# Filter rows for the train and test sets
train_set = combined_df[combined_df['Trip_End'].dt.month.isin(range(1, 12))]  # January to November
test_set = combined_df[combined_df['Trip_End'].dt.month == 12]  # December

# Display the size of each set
print(f"Train set size: {len(train_set)} rows")
print(f"Test set size: {len(test_set)} rows")

train_set.to_csv('../data/processed/2024/train.csv', index=False)
test_set.to_csv('../data/processed/2024/test.csv', index=False)

train_set.head()

from xgboost import XGBRegressor
