==> notebooks/3.IRA_modeling.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_absolute_error, mean_squared_error
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline

df = pd.read_csv("../data/processed/2024/train.csv", parse_dates=['Trip_Start', 'Trip_End'])

df = df.sample(frac=0.1, random_state=42) 

# 1. Calculate trip duration in seconds
df['Trip_Duration'] = (df['Trip_End'] - df['Trip_Start']).dt.total_seconds()


# 2. Extract features from Trip_Start
df['Start_Hour'] = df['Trip_Start'].dt.hour
df['Start_DayOfWeek'] = df['Trip_Start'].dt.dayofweek


# Drop rows with missing or invalid data
df = df.dropna(subset=['Trip_Duration', 'Year_of_Birth', 'Gender', 'Origin_Id', 'Destination_Id'])


# Define features and target variable
X = df[['Year_of_Birth', 'Gender', 'Origin_Id', 'Destination_Id', 'Start_Hour', 'Start_DayOfWeek']]
y = df['Trip_Duration']


# One-hot encode the categorical column 'Gender'
categorical_features = ['Gender']
numerical_features = ['Year_of_Birth', 'Origin_Id', 'Destination_Id', 'Start_Hour', 'Start_DayOfWeek']


# Preprocessing pipeline
preprocessor = ColumnTransformer(
    transformers=[
        ('num', StandardScaler(), numerical_features),
        ('cat', OneHotEncoder(drop='first'), categorical_features)
    ]
)

# Split data into training and testing sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)


X_train.shape

# Baseline Model: Linear Regression
lr_pipeline = Pipeline(steps=[
    ('preprocessor', preprocessor),
    ('regressor', LinearRegression())
])

# Train and evaluate Linear Regression
lr_pipeline.fit(X_train, y_train)
y_pred_lr = lr_pipeline.predict(X_test)


mae_lr = mean_absolute_error(y_test, y_pred_lr)
rmse_lr = np.sqrt(mean_squared_error(y_test, y_pred_lr))

print(f"Linear Regression MAE: {mae_lr}")
print(f"Linear Regression RMSE: {rmse_lr}")

# Advanced Model: Random Forest Regressor
rf_pipeline = Pipeline(steps=[
    ('preprocessor', preprocessor),
    ('regressor', RandomForestRegressor(n_estimators=100, max_depth=3, random_state=42, n_jobs=2))
])

# Train and evaluate Random Forest Regressor
rf_pipeline.fit(X_train, y_train)
y_pred_rf = rf_pipeline.predict(X_test)

mae_rf = mean_absolute_error(y_test, y_pred_rf)
rmse_rf = np.sqrt(mean_squared_error(y_test, y_pred_rf))

print(f"Random Forest Regressor MAE: {mae_rf}")
print(f"Random Forest Regressor RMSE: {rmse_rf}")


import joblib
import datetime

current_date = datetime.datetime.now().strftime("%Y-%m-%d")

model_filename = f"../models/random_forest_mae-{mae_rf:.4f}_{current_date}.pkl"

# Save the trained model to a file
joblib.dump(rf_pipeline, model_filename)

print(f"Model saved successfully as {model_filename}")
